requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
Babel==2.12.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dj-config-url==0.1.1
Django==4.2
django-money==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
py-moneyed==2.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
